wd: v1: fix reserve memory strategy
When SMMU opened, there no need to reserve more memory to
ensure memory is continuous, just allocating the memory
needed by user once.

+++ v1/wd_util.h
@@ -111,7 +111,7 @@ struct q_info {
 	struct wd_ss_region_list ss_list;
 	struct wd_ss_region_list *head;
 	unsigned int dev_flags;
-	unsigned int ss_size;
+	unsigned long ss_size;
 	enum wcrypto_type atype;
 	int ctx_num;
 	struct wd_mm_br br;
